requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.70.0
gunicorn==20.1.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pypika-tortoise==0.1.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==4.0.0
uvicorn==0.15.0
